requirements.txt
alembic==1.4.2
astroid==2.2.5
auth==0.5.3
blinker==1.4
boto==2.49.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.9.2
dnspython==1.16.0
ecdsa==0.13.2
eventlet==0.25.2
falcon==2.0.0
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
future==0.17.1
greenlet==0.4.15
gunicorn==20.0.4
idna==2.9
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.20.0
monotonic==1.5
psycopg2-binary==2.8.2
pyasn1==0.4.8
pycparser==2.20
pylint==2.3.1
pylint-flask-sqlalchemy==0.2.0
pymongo==3.10.1
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.1.0
requests==2.23.0
rsa==4.0
six==1.14.0
SQLAlchemy==1.3.3
typed-ast==1.3.5
urllib3==1.25.9
Werkzeug==0.15.2
wrapt==1.11.1
